refactor: Split reducers
